refactor(error): unify validation errors onto nebula-error canonical type
Three crates each carried their own validation-error world: two types both
named ValidationError (validator with Cow params + pointer-string field;
schema with typed-Value params + segment FieldPath + source chain) plus
expression's ExpressionError. The API layer could only map validator's.
This collapses them onto a single canonical type in the cross-cutting
nebula-error crate.

- nebula-error: new validation module with the canonical ValidationError
  (80-byte boxed-extras layout, RFC 6901 pointer field, severity, params,
  nested, help, plus a source chain), FieldPath, ValidationErrors,
  ValidationReport, and a union builder API (validator new/with_* plus
  convenience ctors AND schema builder().at().param().warn().build()),
  a Classify impl, and serde-gated to_json_value. serde_json added optional
  under the serde feature.
- nebula-validator: deleted its parallel error and field-path modules and
  re-exports the canonical types so crate::foundation::* paths keep
  resolving. Kept codes and ValidationMode.
- nebula-schema: dropped its own ValidationError / Severity /
  ValidationReport, migrated ~110 construction sites, kept the segment
  FieldPath for value-tree navigation, and deleted the dead to_context_map.
- Downstream (credential, credential-runtime, action, api, sdk): updated
  ValidationError access to the canonical surface.

Whole workspace compiles green; nebula-error, validator, schema and the
blast-radius crates test green.
