+++ Cargo.toml
@@ -135,7 +135,11 @@ tower-http = { version = "0.6", features = ["trace", "cors", "compression-gzip"]
 # the `Message::builder()` typed envelope construction;
 # `smtp-transport` brings `AsyncSmtpTransport<Tokio1Executor>`;
 # `tokio1-rustls-tls` is the only TLS stack we ship (matches reqwest above).
-lettre = { version = "0.11", default-features = false, features = ["smtp-transport", "tokio1-rustls-tls", "builder"] }
+lettre = { version = "0.11", default-features = false, features = [
+  "smtp-transport",
+  "tokio1-rustls-tls",
+  "builder",
+] }
 
 # OpenAPI 3.1 spec generation (ADR-0047)
 # `utoipa-axum` is the only mounting path that ties the served `axum::Router`
