+++ crates/schema/src/field.rs
@@ -101,10 +101,6 @@ macro_rules! define_field {
             ///
             /// Returns [`ValidationError`] with code `invalid_key` when `key`
             /// does not satisfy [`FieldKey`] constraints.
-            #[expect(
-                clippy::result_large_err,
-                reason = "ValidationError is intentionally large; callers are on the validation path"
-            )]
             pub fn try_new(key: impl AsRef<str>) -> Result<Self, crate::error::ValidationError> {
                 FieldKey::new(key).map(Self::with_key)
             }
@@ -831,15 +827,10 @@ const FORBIDDEN_EXPRESSION_MODE: ExpressionMode = ExpressionMode::Forbidden;
 // ── Factory methods ───────────────────────────────────────────────────────────
 
 impl Field {
-    #[expect(
-        clippy::result_large_err,
-        reason = "ValidationError is intentionally large; field creation is on the authoring path"
-    )]
     fn parse_key_or_error(key: &str) -> Result<FieldKey, ValidationError> {
         FieldKey::new(key).map_err(|err| {
             ValidationError::builder("invalid_key")
-                .at(FieldPath::root())
-                .param("key", Value::String(key.to_owned()))
+                .param("key", key.to_owned())
                 .message(format!("invalid key `{key}`: {err}"))
                 .build()
         })
@@ -868,7 +859,6 @@ impl Field {
     /// # Errors
     ///
     /// Returns `invalid_key` when `key` cannot be parsed into a [`FieldKey`].
-    #[allow(clippy::result_large_err)]
     pub fn try_string(key: impl AsRef<str>) -> Result<StringField, ValidationError> {
         let key = Self::parse_key_or_error(key.as_ref())?;
         Ok(StringField::with_key(key))
@@ -885,7 +875,6 @@ impl Field {
     /// # Errors
     ///
     /// Returns `invalid_key` when `key` cannot be parsed into a [`FieldKey`].
-    #[allow(clippy::result_large_err)]
     pub fn try_secret(key: impl AsRef<str>) -> Result<SecretField, ValidationError> {
         let key = Self::parse_key_or_error(key.as_ref())?;
         Ok(SecretField::with_key(key))
@@ -902,7 +891,6 @@ impl Field {
     /// # Errors
     ///
     /// Returns `invalid_key` when `key` cannot be parsed into a [`FieldKey`].
-    #[allow(clippy::result_large_err)]
     pub fn try_number(key: impl AsRef<str>) -> Result<NumberField, ValidationError> {
         let key = Self::parse_key_or_error(key.as_ref())?;
         Ok(NumberField::with_key(key))
@@ -919,7 +907,6 @@ impl Field {
     /// # Errors
     ///
     /// Returns `invalid_key` when `key` cannot be parsed into a [`FieldKey`].
-    #[allow(clippy::result_large_err)]
     pub fn try_integer(key: impl AsRef<str>) -> Result<NumberField, ValidationError> {
         let key = Self::parse_key_or_error(key.as_ref())?;
         Ok(NumberField::with_key(key).integer())
@@ -953,7 +940,6 @@ impl Field {
     /// # Errors
     ///
     /// Returns `invalid_key` when `key` cannot be parsed into a [`FieldKey`].
-    #[allow(clippy::result_large_err)]
     pub fn try_boolean(key: impl AsRef<str>) -> Result<BooleanField, ValidationError> {
         let key = Self::parse_key_or_error(key.as_ref())?;
         Ok(BooleanField::with_key(key))
@@ -970,7 +956,6 @@ impl Field {
     /// # Errors
     ///
     /// Returns `invalid_key` when `key` cannot be parsed into a [`FieldKey`].
-    #[allow(clippy::result_large_err)]
     pub fn try_select(key: impl AsRef<str>) -> Result<SelectField, ValidationError> {
         let key = Self::parse_key_or_error(key.as_ref())?;
         Ok(SelectField::with_key(key))
@@ -987,7 +972,6 @@ impl Field {
     /// # Errors
     ///
     /// Returns `invalid_key` when `key` cannot be parsed into a [`FieldKey`].
-    #[allow(clippy::result_large_err)]
     pub fn try_object(key: impl AsRef<str>) -> Result<ObjectField, ValidationError> {
         let key = Self::parse_key_or_error(key.as_ref())?;
         Ok(ObjectField::with_key(key))
@@ -1004,7 +988,6 @@ impl Field {
     /// # Errors
     ///
     /// Returns `invalid_key` when `key` cannot be parsed into a [`FieldKey`].
-    #[allow(clippy::result_large_err)]
     pub fn try_list(key: impl AsRef<str>) -> Result<ListField, ValidationError> {
         let key = Self::parse_key_or_error(key.as_ref())?;
         Ok(ListField::with_key(key))
@@ -1022,7 +1005,6 @@ impl Field {
     /// # Errors
     ///
     /// Returns `invalid_key` when `key` cannot be parsed into a [`FieldKey`].
-    #[allow(clippy::result_large_err)]
     pub fn try_mode(key: impl AsRef<str>) -> Result<ModeField, ValidationError> {
         let key = Self::parse_key_or_error(key.as_ref())?;
         Ok(ModeField::with_key(key))
@@ -1039,7 +1021,6 @@ impl Field {
     /// # Errors
     ///
     /// Returns `invalid_key` when `key` cannot be parsed into a [`FieldKey`].
-    #[allow(clippy::result_large_err)]
     pub fn try_code(key: impl AsRef<str>) -> Result<CodeField, ValidationError> {
         let key = Self::parse_key_or_error(key.as_ref())?;
         Ok(CodeField::with_key(key))
@@ -1056,7 +1037,6 @@ impl Field {
     /// # Errors
     ///
     /// Returns `invalid_key` when `key` cannot be parsed into a [`FieldKey`].
-    #[allow(clippy::result_large_err)]
     pub fn try_file(key: impl AsRef<str>) -> Result<FileField, ValidationError> {
         let key = Self::parse_key_or_error(key.as_ref())?;
         Ok(FileField::with_key(key))
@@ -1073,7 +1053,6 @@ impl Field {
     /// # Errors
     ///
     /// Returns `invalid_key` when `key` cannot be parsed into a [`FieldKey`].
-    #[allow(clippy::result_large_err)]
     pub fn try_computed(key: impl AsRef<str>) -> Result<ComputedField, ValidationError> {
         let key = Self::parse_key_or_error(key.as_ref())?;
         Ok(ComputedField::with_key(key))
@@ -1090,7 +1069,6 @@ impl Field {
     /// # Errors
     ///
     /// Returns `invalid_key` when `key` cannot be parsed into a [`FieldKey`].
-    #[allow(clippy::result_large_err)]
     pub fn try_dynamic(key: impl AsRef<str>) -> Result<DynamicField, ValidationError> {
         let key = Self::parse_key_or_error(key.as_ref())?;
         Ok(DynamicField::with_key(key))
@@ -1107,7 +1085,6 @@ impl Field {
     /// # Errors
     ///
     /// Returns `invalid_key` when `key` cannot be parsed into a [`FieldKey`].
-    #[allow(clippy::result_large_err)]
     pub fn try_notice(key: impl AsRef<str>) -> Result<NoticeField, ValidationError> {
         let key = Self::parse_key_or_error(key.as_ref())?;
         Ok(NoticeField::with_key(key))
