+++ crates/schema/src/prelude.rs
@@ -21,13 +21,14 @@ pub use nebula_schema_macros::EnumSelect;
 pub use nebula_validator::{Predicate, Rule};
 
 pub use crate::{
-    BooleanField, CodeField, ComputedField, ComputedReturn, DynamicField, EvalFuture, Expression,
-    ExpressionContext, ExpressionMode, Field, FieldKey, FieldPath, FieldValue, FieldValues,
-    FileField, HasSchema, HasSelectOptions, InputHint, KdfParams, ListField, LoaderContext,
-    LoaderRegistry, ModeField, ModeVariant, NoticeField, NoticeSeverity, NumberField, ObjectField,
-    RequiredMode, ResolvedValues, Schema, SchemaBuilder, SecretField, SecretValue, SecretWire,
-    SelectField, SelectOption, Severity, StringField, Transformer, ValidSchema, ValidValues,
-    ValidationError, ValidationReport, VisibilityMode, builder::FieldCollector, field_key,
+    BooleanField, CodeField, ComputedField, ComputedReturn, DynamicField, ErrorSeverity,
+    EvalFuture, Expression, ExpressionContext, ExpressionMode, Field, FieldKey, FieldPath,
+    FieldValue, FieldValues, FileField, HasSchema, HasSelectOptions, InputHint, KdfParams,
+    ListField, LoaderContext, LoaderRegistry, ModeField, ModeVariant, NoticeField, NoticeSeverity,
+    NumberField, ObjectField, RequiredMode, ResolvedValues, Schema, SchemaBuilder, SecretField,
+    SecretValue, SecretWire, SelectField, SelectOption, StringField, Transformer, ValidSchema,
+    ValidValues, ValidationError, ValidationReport, VisibilityMode, builder::FieldCollector,
+    field_key,
 };
 
 #[cfg(test)]
